+++ src/app.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Route } from "react-router-dom";
 import { useAuth0 } from "@auth0/auth0-react";
-import "./app.css";
+// import "./app.css";
 import LoginView from "./views/loginView";
 // import Home from "./views/homeView";
 import SelectOrginization from "./views/selectOrganizationView";