+++ src/views/editStopView.js
@@ -1,17 +1,15 @@
 //component for the edit stop view
 //access through the EditTour component
 
-import React, { Fragment, useState} from "react";
-import { render } from "react-dom";
-
+import React from "react";
+import Button from "react-bootstrap/Button";
 //import the css module
-import styles from "../css_modules/generalStyles.module.css";
-import EditTour from "./editTourView";
+
 
 const EditStop = ({setEditStopMode}) => {
 
     return (
-        <div className={styles.generalStyle}>
+        <div>
             <h1>This is the edit stop view</h1>
             <p>
             stuff to do:
@@ -20,7 +18,7 @@ const EditStop = ({setEditStopMode}) => {
                 <li>get google maps plugin</li>
             </ul>
             </p>
-            <button onClick={() => goBackToTourEdit()}>go back</button>
+            <Button onClick={() => goBackToTourEdit()}>go back</Button>
         </div>
     );